tests/wpt/web-platform-tests/css/css-writing-modes/parsing/text-orientation-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Writing Modes: getComputedStyle().textOrientation</title>
<link rel="help" href="https://drafts.csswg.org/css-writing-modes-3/#propdef-text-orientation">
<meta name="assert" content="text-orientation computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-orientation", "mixed");
test_computed_value("text-orientation", "upright");
test_computed_value("text-orientation", "sideways");
</script>
</body>
</html>
